+++ src/biz/bokhorst/xprivacy/XFusedLocationApi.java
@@ -46,7 +46,7 @@ public class XFusedLocationApi extends XHook {
 
 	public static List<XHook> getInstances(Object instance) {
 		String className = instance.getClass().getName();
-		Util.log(null, Log.INFO, "Hooking class=" + className + " uid=" + Binder.getCallingUid());
+		Util.log(null, Log.INFO, "Hooking FusedLocationApi class=" + className + " uid=" + Binder.getCallingUid());
 
 		List<XHook> listHook = new ArrayList<XHook>();
 		listHook.add(new XFusedLocationApi(Methods.getLastLocation, PrivacyManager.cLocation, className));